Overhaul audio processor to provide HTTP API
This fits better with cloud GPU PaaS offerings, because exposing the
redis queue directly adds security and networking complexity. I'd like
to eventually implement a WebSocket pull operation mode as well, so
workers don't need to be listening on the network.

A happy side-effect is no longer needing a doc worker: we can do the
document updates inline with consuming the queue in Deno-land.