Fix punctuation / whitespace splitting on paste
This adds a heuristic to ensure punctuation before / after a pasted word
from another speaker gets split to the right line. It's not entirely
perfect, though, e.g. pasting a word right before a ? steals that
character. It'd be convenient if it was possible to detect if the node
was just pasted -- isDirty() doesn't seem to do it.