More app coalesce-shared fixups
Digging deeper into the hole of hacking shared dependency resolution via
aliases and import maps. An import-map will be necessary for sharing
with Deno, anyway, but I'd hoped npm would install the local package
sub-dependencies. It looks like this might be possible with npm's
`install-links=true` option or pnpm, but for now this is the quickest
fix.